server-src: Latest server with all of the end points.
