+++ server/src/ratings.py
@@ -168,107 +168,118 @@ def getTasteRecipes(user_id, recipes_wanted):
 # updateSingleIngredientRating
 # Updates the user's rating of the ingredient, ingredient subcluster and cluster.
 # - Input:
-#   - (string) user_id,
+#   - (document) user_doc,
 #   - (string) ingredient_id,
-#   - (float)  rating,
-#   - (string) rating_type
+#   - (dict)  ratings,
+#   - (array - string) rating_types
 # - Output:
+#   - (dict) update_dict,
 #   - (string) error
-def updateSingleIngredientRating(user_id, ingredient_id, rating, rating_type):
-  rating_type = rating_type.lower()
-  debug(f'[updateSingleIngredientRating - {rating_type} - INFO]: Starting.')
+def updateSingleIngredientRating(user_doc, ingredient_id, ratings, rating_types):
+  debug(f'[updateSingleIngredientRating - INFO]: Starting.')
   updating_data = {}
 
-  if not (rating_type in rating_types):
-    err = f'[updateSingleIngredientRating - {rating_type} - ERROR]: rating_type {rating_type} is not a known rating type.'
-    debug(err)
-    return err
+  for rating_type in rating_types:
+    if not (rating_type in rating_types):
+      err = f'[updateSingleIngredientRating - ERROR]: rating_type {rating_type} is not a known rating type.'
+      debug(err)
+      return None, err
 
   # Retrieve the user document
-  doc_ref, doc, err = retrieveDocument('users', user_id)
-  if err:
-    return err
-  user_dict = doc.to_dict()
+  user_dict = user_doc.to_dict()
 
   # Retrieve the ingredients document
   doc_ref, doc, err = retrieveDocument('ingredients', ingredient_id)
   if err:
-    err = '[updateSingleIngredientRating - {rating_type} - ERROR]: Unable to obtain ingredient {ingredient_id} information from DB.'
+    err = '[updateSingleIngredientRating - ERROR]: Unable to obtain ingredient {ingredient_id} information from DB.'
     debug(err)
-    return err
+    return None, err
   ingredients_dict = doc.to_dict()
 
   # Update the ingredient rating
-  updating_data['i_'+rating_type] = user_dict['i_'+rating_type]
-  try:
-    r = user_dict['i_'+rating_type][ingredient_id]['rating']
-    n = user_dict['i_'+rating_type][ingredient_id]['n_ratings']
-    r = (r*n+rating)/(n+1)
-    n += 1
-    updating_data['i_'+rating_type][ingredient_id] =  {'rating': r, 'n_ratings': n}
-  except:
-    updating_data['i_'+rating_type][ingredient_id] = {'rating': rating, 'n_ratings': 1}
-
-  # Update the ingredient subcluster rating
-  subcluster_id = str(ingredients_dict["subcluster"])
-  if subcluster_id != None:
-    updating_data['is_'+rating_type] = user_dict['is_'+rating_type]
+  for rating_type in rating_types:
+    rating = ratings[rating_type]
+    updating_data['i_'+rating_type] = user_dict['i_'+rating_type]
     try:
-      r = user_dict['is_'+rating_type][subcluster_id]['rating']
-      n = user_dict['is_'+rating_type][subcluster_id]['n_ratings']
+      r = user_dict['i_'+rating_type][ingredient_id]['rating']
+      n = user_dict['i_'+rating_type][ingredient_id]['n_ratings']
       r = (r*n+rating)/(n+1)
       n += 1
-      updating_data['is_'+rating_type][subcluster_id] = {'rating': r, 'n_ratings': n}
+      updating_data['i_'+rating_type][ingredient_id] =  {'rating': r, 'n_ratings': n}
     except:
-      updating_data['is_'+rating_type][subcluster_id] = {'rating': rating, 'n_ratings': 1}
+      updating_data['i_'+rating_type][ingredient_id] = {'rating': rating, 'n_ratings': 1}
 
-  # Update the ingredient cluster rating
-  cluster_id = str(ingredients_dict["cluster"])
-  if cluster_id != None:
-    updating_data['ic_'+rating_type] = user_dict['ic_'+rating_type]
-    try:
-      r = user_dict['ic_'+rating_type][cluster_id]['rating']
-      n = user_dict['ic_'+rating_type][cluster_id]['n_ratings']
-      r = (r*n+rating)/(n+1)
-      n += 1
-      updating_data['ic_'+rating_type][cluster_id] = {'rating': r, 'n_ratings': n}
-    except:
-      updating_data['ic_'+rating_type][cluster_id] = {'rating': rating, 'n_ratings': 1}
+    # Update the ingredient subcluster rating
+    subcluster_id = str(ingredients_dict["subcluster"])
+    if subcluster_id != None:
+      updating_data['is_'+rating_type] = user_dict['is_'+rating_type]
+      try:
+        r = user_dict['is_'+rating_type][subcluster_id]['rating']
+        n = user_dict['is_'+rating_type][subcluster_id]['n_ratings']
+        r = (r*n+rating)/(n+1)
+        n += 1
+        updating_data['is_'+rating_type][subcluster_id] = {'rating': r, 'n_ratings': n}
+      except:
+        updating_data['is_'+rating_type][subcluster_id] = {'rating': rating, 'n_ratings': 1}
+
+    # Update the ingredient cluster rating
+    cluster_id = str(ingredients_dict["cluster"])
+    if cluster_id != None:
+      updating_data['ic_'+rating_type] = user_dict['ic_'+rating_type]
+      try:
+        r = user_dict['ic_'+rating_type][cluster_id]['rating']
+        n = user_dict['ic_'+rating_type][cluster_id]['n_ratings']
+        r = (r*n+rating)/(n+1)
+        n += 1
+        updating_data['ic_'+rating_type][cluster_id] = {'rating': r, 'n_ratings': n}
+      except:
+        updating_data['ic_'+rating_type][cluster_id] = {'rating': rating, 'n_ratings': 1}
 
   # Update the user's document:
-  err = updateDocument('users', user_id, updating_data)
-  if err:
-    err = f'[updateSingleIngredientRating - {rating_type} - ERROR]: Unable to update user document with ratings for ingredient {ingredient_id}, err: {err}'
-    debug(err)
-    return err
-  return ''
+  return updating_data, ''
 
 ################################################################################
 # updateIngredientRating
 # Updates the user's rating of the given ingredients.
 # - Input:
 #   - (dict) data containing user id, ingredient ids and ratings,
-#   - (string) rating_type
+#   - (array - string) rating_type
 # - Output:
 #   - (string) error
-def updateIngredientRatings(data, rating_type):
-  rating_type = rating_type.lower()
-  debug(f'[updateIngredientRating - {rating_type} - INFO]: Starting.')
-
-  if not (rating_type in rating_types):
-    err = f'[updateIngredientRating - {rating_type} - ERROR]: rating_type {rating_type} is not a known rating type.'
-    debug(err)
-    return err
-
+def updateIngredientRatings(data, rating_types):
+  debug(f'[updateIngredientRating - INFO]: Starting.')
+  updating_data = {}
   user_id = data['userID']
-  ingredient_ids = data[rating_type+'_ratings']
-  for ingredient_id in ingredient_ids.keys():
-    rating = ingredient_ids[ingredient_id]
-    err = updateSingleIngredientRating(user_id, ingredient_id, rating, rating_type)
-    if err:
-      err = f'[updateIngredientRating - {rating_type} - ERROR]: Unable to update ratings for ingredient {ingredient_id}, err: {err}'
+
+  for rating_type in rating_types:
+    if not (rating_type in rating_types):
+      err = f'[updateIngredientRating - ERROR]: rating_type {rating_type} is not a known rating type.'
       debug(err)
       return err
+
+  # Retrieve the user document
+  user_doc_ref, user_doc, err = retrieveDocument('users', user_id)
+  if err:
+    return err
+
+  ingredient_ids = data[rating_types[0]]
+  for ingredient_id in ingredient_ids.keys():
+    ratings = {}
+    for rating_type in rating_types:
+      ratings[rating_type] = data[rating_type+'_ratings'][ingredient_id]
+    update_dict, err = updateSingleIngredientRating(user_doc, ingredient_id, ratings, rating_types)
+    if err:
+      err = f'[updateIngredientRating - ERROR]: Unable to update ratings for ingredient {ingredient_id}, err: {err}'
+      debug(err)
+      return err
+    updating_data.update(update_dict)
+
+  # Update the user's document:
+  err = updateDocument('users', user_id, updating_data)
+  if err:
+    err = f'[updateIngredientRating - ERROR]: Unable to update user document with ratings, err: {err}'
+    debug(err)
+    return err
   return ''
 
 ################################################################################
@@ -276,21 +287,20 @@ def updateIngredientRatings(data, rating_type):
 # Updates the user's rating of the given ingredient clusters.
 # - Input:
 #   - (dict) data containing user id, ingredient cluster ids and ratings,
-#   - (string) rating_type
+#   - (array - string) rating_types
 # - Output:
 #   - (string) error
-def updateIngredientClusterRatings(data, rating_type):
-  rating_type = rating_type.lower()
-  debug(f'[updateIngredientClusterRatings - {rating_type} - INFO]: Starting.')
-  updating_data = {'ic_'+rating_type: {}}
+def updateIngredientClusterRatings(data, rating_types):
+  debug(f'[updateIngredientClusterRatings - INFO]: Starting.')
+  updating_data = {}
   user_id = data['userID']
-  ic_ids = data[rating_type+'_ratings']
-  print(ic_ids)
 
-  if not (rating_type in rating_types):
-    err = f'[updateIngredientClusterRatings - {rating_type} - ERROR]: rating_type {rating_type} is not a known rating type.'
-    debug(err)
-    return err
+  for rating_type in rating_types:
+    if not (rating_type in rating_types):
+      err = f'[updateIngredientClusterRatings - ERROR]: rating_type {rating_type} is not a known rating type.'
+      debug(err)
+      return err
+    updating_data['ic_'+rating_type] = {}
 
   # Retrieve the user document
   doc_ref, doc, err = retrieveDocument('users', user_id)
@@ -299,26 +309,23 @@ def updateIngredientClusterRatings(data, rating_type):
   user_dict = doc.to_dict()
 
   # Update the ic values
-  updating_data['ic_'+rating_type] = user_dict['ic_'+rating_type]
-  print(updating_data)
+  ic_ids = data[rating_types[0]+'_ratings']
   for ic_id in ic_ids.keys():
-    print(ic_id)
-    rating = ic_ids[ic_id]
-    try:
-      r = user_dict['ic_'+rating_type][ic_id]['rating']
-      n = user_dict['ic_'+rating_type][ic_id]['n_ratings']
-      r = (r*n+rating)/(n+1)
-      n += 1
-      updating_data['ic_'+rating_type][ic_id] = {'rating': r, 'n_ratings': n}
-    except:
-      updating_data['ic_'+rating_type][ic_id] = {'rating': rating, 'n_ratings': 1}
-    print(updating_data)
+    for rating_type in rating_types:
+      rating = data[rating_type+'_ratings'][ic_id]
+      try:
+        r = user_dict['ic_'+rating_type][ic_id]['rating']
+        n = user_dict['ic_'+rating_type][ic_id]['n_ratings']
+        r = (r*n+rating)/(n+1)
+        n += 1
+        updating_data['ic_'+rating_type][ic_id] = {'rating': r, 'n_ratings': n}
+      except:
+        updating_data['ic_'+rating_type][ic_id] = {'rating': rating, 'n_ratings': 1}
 
   # Update the user document
-  print(updating_data)
   err = updateDocument('users', user_id, updating_data)
   if err:
-    err = f'[updateIngredientClusterRatings - {rating_type} - ERROR]: Unable to update user document with ratings for ingredient clusters rated, err: {err}'
+    err = f'[updateIngredientClusterRatings - ERROR]: Unable to update user document with ratings for ingredient clusters rated, err: {err}'
     debug(err)
     return err
   return ''
@@ -327,27 +334,26 @@ def updateIngredientClusterRatings(data, rating_type):
 # updateSingleRecipeRating
 # Updates the user's rating of the recipe and it's ingredients.
 # - Input:
-#   - (string) user id,
-#   - (string) recipe id
-#   - (float)  rating,
-#   - (string) rating_type
+#   - (document) user_doc,
+#   - (string) recipe_id
+#   - (dict)  ratings,
+#   - (array - string) rating_types
 # - Output:
+#   - (dict) update_dict,
 #   - (string) error
-def updateSingleRecipeRating(user_id, recipe_id, rating, rating_type):
-  rating_type = rating_type.lower()
-  debug(f'[updateSingleRecipeRating - {rating_type} - INFO]: Starting.')
-  updating_data = {'r_'+rating_type: {}}
+def updateSingleRecipeRating(user_doc, recipe_id, ratings, rating_types):
+  debug(f'[updateSingleRecipeRating - INFO]: Starting.')
+  updating_data = {}
 
-  if not (rating_type in rating_types):
-    err = f'[updateSingleRecipeRating - {rating_type} - ERROR]: rating_type {rating_type} is not a known rating type.'
-    debug(err)
-    return err
+  for rating_type in rating_types:
+    if not (rating_type in rating_types):
+      err = f'[updateSingleRecipeRating - ERROR]: rating_type {rating_type} is not a known rating type.'
+      debug(err)
+      return err
+    updating_data['ic_'+rating_type] = {}
 
   # Retrieve the user document
-  doc_ref, doc, err = retrieveDocument('users', user_id)
-  if err:
-    return err
-  user_dict = doc.to_dict()
+  user_dict = user_doc.to_dict()
 
   # Retrieve the recipe document
   doc_ref, doc, err = retrieveDocument('recipes', recipe_id)
@@ -355,59 +361,73 @@ def updateSingleRecipeRating(user_id, recipe_id, rating, rating_type):
     return err
   recipe_dict = doc.to_dict()
 
-  # Update the recipe rating
-  updating_data['r_'+rating_type] = user_dict['r_'+rating_type]
-  try:
-    r = user_dict['r_'+rating_type][recipe_id]['rating']
-    n = user_dict['r_'+rating_type][recipe_id]['n_ratings']
-    r = (r*n+rating)/(n+1)
-    n += 1
-    updating_data['r_'+rating_type][recipe_id] =  {'rating': r, 'n_ratings': n}
-  except:
-    updating_data['r_'+rating_type][recipe_id] = {'rating': rating, 'n_ratings': 1}
+  # Update the recipe ratings
+  for rating_type in rating_types:
+    updating_data['r_'+rating_type] = user_dict['r_'+rating_type]
+    rating = ratings[rating_type]
+    try:
+      r = user_dict['r_'+rating_type][recipe_id]['rating']
+      n = user_dict['r_'+rating_type][recipe_id]['n_ratings']
+      r = (r*n+rating)/(n+1)
+      n += 1
+      updating_data['r_'+rating_type][recipe_id] =  {'rating': r, 'n_ratings': n}
+    except:
+      updating_data['r_'+rating_type][recipe_id] = {'rating': rating, 'n_ratings': 1}
 
   # Update the ingredients.
   ingredient_ids = recipe_dict["ingredient_ids"]
   for ingredient_id in ingredient_ids:
     ingredient_id = str(ingredient_id)
-    err = updateSingleIngredientRating(user_id, ingredient_id, rating, rating_type)
+    update_dict, err = updateSingleIngredientRating(user_doc, ingredient_id, ratings, rating_types)
     if err:
-      err = f'[updateSingleRecipeRating - {rating_type} - ERROR]: Unable to update ratings for recipe {recipe_id}, err: {err}'
+      err = f'[updateSingleRecipeRating - ERROR]: Unable to update ratings for recipe {recipe_id}, err: {err}'
       debug(err)
       return err
+    updating_data.update(update_dict)
 
   # Update the user's document:
-  err = updateDocument('users', user_id, updating_data)
-  if err:
-    err = f'[updateSingleRecipeRating - {rating_type} - ERROR]: Unable to update user document with ratings for recipe {recipe_id}, err: {err}'
-    debug(err)
-    return err
-  return''
+  return updating_data, ''
 
 ################################################################################
 # updateRecipeRatings
 # Updates the user's recipe rating of the given recipes.
 # - Input:
 #   - (dict) data containing user id, ingredient ids and ratings,
-#   - (string) rating_type
+#   - (array - string) rating_types
 # - Output:
 #   - (string) error
-def updateRecipeRatings(data, rating_type):
-  rating_type = rating_type.lower()
-  debug(f'[updateRecipeRatings - {rating_type} - INFO]: Starting.')
-
-  if not (rating_type in rating_types):
-    err = f'[updateRecipeRatings - {rating_type} - ERROR]: rating_type {rating_type} is not a known rating type.'
-    debug(err)
-    return err
-
+def updateRecipeRatings(data, rating_types):
+  debug(f'[updateRecipeRatings - INFO]: Starting.')
+  updating_data = {}
   user_id = data['userID']
-  recipe_ids = data[rating_type+'_ratings']
-  for recipe_id in recipe_ids.keys():
-    rating = recipe_ids[recipe_id]
-    err = updateSingleRecipeRating(user_id, recipe_id, rating, rating_type)
-    if err:
-      err = f'[updateRecipeRatings - {rating_type} - ERROR]: Unable to update ratings for recipe {recipe_id}, err: {err}'
+
+  for rating_type in rating_types:
+    if not (rating_type in rating_types):
+      err = f'[updateRecipeRatings - ERROR]: rating_type {rating_type} is not a known rating type.'
       debug(err)
       return err
+
+  # Retrieve the user document
+  user_doc_ref, user_doc, err = retrieveDocument('users', user_id)
+  if err:
+    return err
+
+  recipe_ids = data[rating_types[0]+'_ratings']
+  for recipe_id in recipe_ids.keys():
+    ratings = {}
+    for rating_type in rating_types:
+      ratings[rating_type] = data[rating_type+'_ratings'][recipe_id]
+    update_dict, err = updateSingleRecipeRating(user_doc, recipe_id, ratings, rating_types)
+    if err:
+      err = f'[updateRecipeRatings - ERROR]: Unable to update ratings for recipe {recipe_id}, err: {err}'
+      debug(err)
+      return err
+    updating_data.update(update_dict)
+
+  # Update the user document
+  err = updateDocument('users', user_id, updating_data)
+  if err:
+    err = f'[updateRecipeRatings - ERROR]: Unable to update user document with ratings for recipes and ingredients, err: {err}'
+    debug(err)
+    return err
   return ''
